+++ toml/toml_test.go
@@ -2,6 +2,7 @@ package toml_test
 
 import (
 	"bytes"
+	"fmt"
 	"strings"
 	"testing"
 	"time"
@@ -11,23 +12,54 @@ import (
 	itoml "github.com/influxdata/influxdb/toml"
 )
 
-// Ensure that megabyte sizes can be parsed.
-func TestSize_UnmarshalText_MB(t *testing.T) {
-	var s itoml.Size
-	if err := s.UnmarshalText([]byte("200m")); err != nil {
-		t.Fatalf("unexpected error: %s", err)
-	} else if s != 200*(1<<20) {
-		t.Fatalf("unexpected size: %d", s)
-	}
-}
+func TestSize_UnmarshalText(t *testing.T) {
+	// Copy this from the toml package
+	maxInt := int64(^uint(0) >> 1)
 
-// Ensure that gigabyte sizes can be parsed.
-func TestSize_UnmarshalText_GB(t *testing.T) {
 	var s itoml.Size
-	if err := s.UnmarshalText([]byte("1g")); err != nil {
-		t.Fatalf("unexpected error: %s", err)
-	} else if s != 1073741824 {
-		t.Fatalf("unexpected size: %d", s)
+	for _, test := range []struct {
+		str  string
+		want int64
+	}{
+		{"1", 1},
+		{"10", 10},
+		{"100", 100},
+		{"1k", 1 << 10},
+		{"10k", 10 << 10},
+		{"100k", 100 << 10},
+		{"1K", 1 << 10},
+		{"10K", 10 << 10},
+		{"100K", 100 << 10},
+		{"1m", 1 << 20},
+		{"10m", 10 << 20},
+		{"100m", 100 << 20},
+		{"1M", 1 << 20},
+		{"10M", 10 << 20},
+		{"100M", 100 << 20},
+		{"1g", 1 << 30},
+		{"1G", 1 << 30},
+		{fmt.Sprint(maxInt - 1), maxInt - 1},
+	} {
+		if err := s.UnmarshalText([]byte(test.str)); err != nil {
+			t.Fatalf("unexpected error: %s", err)
+		}
+		if s != itoml.Size(test.want) {
+			t.Fatalf("wanted: %d got: %d", test.want, s)
+		}
+	}
+
+	for _, str := range []string{
+		fmt.Sprintf("%dk", maxInt-1),
+		"10000000000000000000g",
+		"abcdef",
+		"1KB",
+		"√m",
+		"a1",
+		"",
+	} {
+		if err := s.UnmarshalText([]byte(str)); err == nil {
+			t.Fatalf("input should have failed: %s", str)
+		}
 	}
 }